requirements.txt
Flask==3.0.0
google-cloud-storage==2.14.0
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
